organizando los estilos
